Revert "Added tabs. Moved initiative to a page and created SpellControl. Routes"
