+++ src/app/app-routing.module.ts
@@ -8,7 +8,7 @@ const routes: Routes = [
   },
   {
     path: '',
-    redirectTo: 'home/initiative',
+    redirectTo: 'home',
     pathMatch: 'full'
   },
 ];
