--- src/app/pages/initiative/initiative-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { InitiativePage } from './initiative.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: InitiativePage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class InitiativePageRoutingModule {}
