--- src/app/pages/initiative/initiative.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { IonicModule } from '@ionic/angular';
-
-import { InitiativePageRoutingModule } from './initiative-routing.module';
-
-import { InitiativePage } from './initiative.page';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    InitiativePageRoutingModule
-  ],
-  declarations: [InitiativePage]
-})
-export class InitiativePageModule {}
