--- src/app/pages/spellscontrol/spellscontrol.page.html
@@ -1,25 +1,0 @@
-<ion-content>
-  <div class="app-title">⭐ Spells Control</div>
-  <div style="width: 80%; margin: auto; color: white;">
-    <ion-list *ngFor="let entity of entityArray; let i = index">
-      <ion-item style="color: white;">
-        <ion-label (click)="editEntity(entity, i)">
-          <span class="initiative">{{ entity.initiative }}</span>{{ entity.name }}
-        </ion-label>
-        <ion-button fill="clear" color="danger" (click)="deleteEntity(i)">
-          <ion-icon slot="icon-only" name="trash-outline"></ion-icon>
-        </ion-button>
-      </ion-item>
-    </ion-list>
-    <div style="margin: 20px auto auto auto">
-      <ion-button (click)="addEntity()">Add entity</ion-button>
-      <ion-button color="success" (click)="roll(4)">D4</ion-button>
-      <ion-button color="success" (click)="roll(6)">D6</ion-button>
-      <ion-button color="success" (click)="roll(8)">D8</ion-button>
-      <ion-button color="success" (click)="roll(10)">D10</ion-button>
-      <ion-button color="success" (click)="roll(12)">D12</ion-button>
-      <ion-button color="success" (click)="roll(20)">D20</ion-button>
-      <ion-button color="success" (click)="roll(100)">D100</ion-button>
-    </div>
-  </div>
-</ion-content>
