--- src/app/pages/spellscontrol/spellscontrol.page.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-spellscontrol',
-  templateUrl: './spellscontrol.page.html',
-  styleUrls: [
-    './spellscontrol.page.scss',
-    '../../app.component.scss'
-  ],
-})
-export class SpellscontrolPage implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
